dummy data completed

+++ db/helpers/manage-tables.js
@@ -1,7 +1,7 @@
-const db = require('../connection');
+const db = require("../connection");
 
 const createTables = async () => {
-	const townsTablePromise = db.query(`
+  const townsTablePromise = db.query(`
   CREATE TABLE towns (
     town_id SERIAL PRIMARY KEY,
     town_name VARCHAR NOT NULL,
@@ -9,7 +9,7 @@ const createTables = async () => {
 
   );`);
 
-	const usersTablePromise = db.query(`
+  const usersTablePromise = db.query(`
   CREATE TABLE users (
     user_id SERIAL PRIMARY KEY,
     username VARCHAR NOT NULL,
@@ -18,9 +18,9 @@ const createTables = async () => {
 
   );`);
 
-	await Promsise.all([townsTablePromise, usersTablePromise]);
+  await Promsise.all([townsTablePromise, usersTablePromise]);
 
-	await db.query(`
+  await db.query(`
   CREATE TABLE parks (
     park_id SERIAL PRIMARY KEY,
     town_id INT NOT NULL REFERENCES towns(town_id),
@@ -30,20 +30,20 @@ const createTables = async () => {
     amenities VARCHAR NOT NULL
   );`);
 
-	//Count for total ammount of hunts on parks
+  //Count for total ammount of hunts on parks
 
-	await db.query(`
+  await db.query(`
   CREATE TABLE maps (
     map_id SERIAL PRIMARY KEY,
     park_id INT NOT NULL REFERENCES parks(park_id),
-    length INT NOT NULL,
+    length DECIMAL NOT NULL,
     est_comp_time INT NOT NULL,
     age_min INT NOT NULL
   );`);
 
-	//Count for total ammount of waypoints on maps
+  //Count for total ammount of waypoints on maps
 
-	await db.query(`
+  await db.query(`
   CREATE TABLE user_activity (
     activity_id SERIAL PRIMARY KEY,
     user_id INT NOT NULL REFERENCES users(user_id),
@@ -54,11 +54,11 @@ const createTables = async () => {
 };
 
 const dropTables = async () => {
-	await db.query(`DROP TABLE IF EXISTS towns`);
-	await db.query(`DROP TABLE IF EXISTS users`);
-	await db.query(`DROP TABLE IF EXISTS parks`);
-	await db.query(`DROP TABLE IF EXISTS maps`);
-	await db.query(`DROP TABLE IF EXISTS user_activity`);
+  await db.query(`DROP TABLE IF EXISTS towns`);
+  await db.query(`DROP TABLE IF EXISTS users`);
+  await db.query(`DROP TABLE IF EXISTS parks`);
+  await db.query(`DROP TABLE IF EXISTS maps`);
+  await db.query(`DROP TABLE IF EXISTS user_activity`);
 };
 
 module.exports = { createTables, dropTables };
